+++ SimpleTcp/ClientConnectedEventArgs.cs
@@ -2,13 +2,19 @@
 
 namespace SimpleTcp
 {
+    /// <summary>
+    /// Arguments for client connection events.
+    /// </summary>
     public class ClientConnectedEventArgs : EventArgs
     {
-        internal ClientConnectedEventArgs(string ipAndPort)
+        internal ClientConnectedEventArgs(string ipPort)
         {
-            IpAndPort = ipAndPort;
+            IpPort = ipPort;
         }
 
-        public string IpAndPort { get; }
+        /// <summary>
+        /// The IP address and port number of the connected client socket.
+        /// </summary>
+        public string IpPort { get; }
     }
 }